Enhance current player label.
I now show the current player's color as the background color of the
label. I also made changes to how it is laid out in the frame. I use a
box layout for the status bar panel instead of a flow layout. I also
changed how I set the label, adding setCurrentPlayer() to the view.