Add checkmate detection to controller.
The game can now recognize when a player reaches checkmate. Yay!
Although it will require futher testing to make sure it's working
properly. I added a setWinner method to the view which I set in the
controller after checking for checkmate/stalemate.

Note: I think there is a problem with the highlighting functionality
because it sometimes doesn't disappear when it's supposed to.