Add canCastle() to ChessBoard.
This is a big method because there's a lot of conditions required for
castling, but its flat so it should be readable. I included tests for
it. I think I will rename my unit tests with this format later:
methodName_longDescription

I think it makes it easier to determine where the method name ends with
the underscore added.