Replace atValidStartingPosition with inStartingPosition.
I only care about something being in its starting position if it hasn't
moved yet. I made hasMoved protected and added a check for it to the
beginning of every inStartingPosition definition.  I renamed the method
to eliminate the three-way distinction between valid starting positions,
invalid starting positions, and non-starting positions. Now it's either
in its starting position or it's not.