Make move() work like forceMove().
I pretty much always want to check if something is a valid move before I
call move. Putting the isValidMove() check in move caused some problems,
necessitating the creation of forceMove(). Will remove forceMove in a
later commit now that it is redundant. I had to remove some poorly
written tests that were nothing but a list of asserts on moves.