Add en passant to game.
This gives pawns the ability to perform an en passant capture. It
contains all necessary functionality. canEnPassant and enPassant methods
were added to ChessBoard (with tests). The ability to perform an en
passant in game was added to the controller along with a small change to
the view that allows a user to clear a chosen space.